basic action implemented, not tested
TODO state rewrite
