+++ app/cleverobot/templates/index.html
@@ -8,7 +8,7 @@
 {% endblock %}
 {% block content %}
 <form action="">
-    <input id="m" autocomplete="off" value="Hello..."/>
+    <input id="m" autocomplete="off" value="Hello"/>
     <button>Send</button>
 </form>
 <ul id="messages">
