+++ cbot/kb/kb_data.py
@@ -1,0 +1,366 @@
+#!/usr/bin/env python
+# encoding: utf-8
+from __future__ import unicode_literals
+
+
+data = {}
+
+bot_properties = {
+    ('cleverobot', 'known_as', 'cleverobot'),
+    ('cleverobot', 'known_as', 'I'),
+}
+
+is_capital = {
+    ('Montgomery', 'isCapitalOf', 'Alabama'),
+    ('Juneau', 'isCapitalOf', 'Alaska'),
+    ('Phoenix', 'isCapitalOf', 'Arizona'),
+    ('Little Rock', 'isCapitalOf', 'Arkansas'),
+    ('Sacramento', 'isCapitalOf', 'California'),
+    ('Denver', 'isCapitalOf', 'Colorado'),
+    ('Hartford', 'isCapitalOf', 'Connecticut'),
+    ('Dover', 'isCapitalOf', 'Delaware'),
+    ('Tallahassee', 'isCapitalOf', 'Florida'),
+    ('Honolulu', 'isCapitalOf', 'Hawaii'),
+    ('Boise', 'isCapitalOf', 'Idaho'),
+    ('Springfield', 'isCapitalOf', 'Illinois'),
+    ('Des Moines', 'isCapitalOf', 'Iowa'),
+    ('Topeka', 'isCapitalOf', 'Kansas'),
+    ('Frankfort', 'isCapitalOf', 'Kentucky'),
+    ('Baton Rouge', 'isCapitalOf', 'Louisiana'),
+    ('Augusta', 'isCapitalOf', 'Maine'),
+    ('Annapolis', 'isCapitalOf', 'Maryland'),
+    ('Lansing', 'isCapitalOf', 'Michigan'),
+    ('Saint Paul', 'isCapitalOf', 'Minnesota'),
+    ('Jackson', 'isCapitalOf', 'Mississippi'),
+    ('Jefferson City', 'isCapitalOf', 'Missouri'),
+    ('Helena', 'isCapitalOf', 'Montana'),
+    ('Lincoln', 'isCapitalOf', 'Nebraska'),
+    ('Carson City', 'isCapitalOf', 'Nevada'),
+    ('Concord', 'isCapitalOf', 'New Hampshire'),
+    ('Trenton', 'isCapitalOf', 'New Jersey'),
+    ('Santa Fe', 'isCapitalOf', 'New Mexico'),
+    ('Albany', 'isCapitalOf', 'New York'),
+    ('Raleigh', 'isCapitalOf', 'North Carolina'),
+    ('Bismarck', 'isCapitalOf', 'North Dakota'),
+    ('Columbus', 'isCapitalOf', 'Ohio'),
+    ('Salem', 'isCapitalOf', 'Oregon'),
+    ('Harrisburg', 'isCapitalOf', 'Pennsylvania'),
+    ('Providence', 'isCapitalOf', 'Rhode Island'),
+    ('Columbia', 'isCapitalOf', 'South Carolina'),
+    ('Pierre', 'isCapitalOf', 'South Dakota'),
+    ('Nashville', 'isCapitalOf', 'Tennessee'),
+    ('Austin', 'isCapitalOf', 'Texas'),
+    ('Montpelier', 'isCapitalOf', 'Vermont'),
+    ('Richmond', 'isCapitalOf', 'Virginia'),
+    ('Charleston', 'isCapitalOf', 'West Virginia'),
+    ('Madison', 'isCapitalOf', 'Wisconsin'),
+    ('Cheyenne', 'isCapitalOf', 'Wyoming'),
+    ('Charlotte Amalie', 'isCapitalOf', 'United States Virgin Islands'),
+}
+
+is_city = {
+    ('Montgomery', 'isA', 'City'),
+    ('Juneau', 'isA', 'City'),
+    ('Phoenix', 'isA', 'City'),
+    ('Little Rock', 'isA', 'City'),
+    ('Sacramento', 'isA', 'City'),
+    ('Denver', 'isA', 'City'),
+    ('Hartford', 'isA', 'City'),
+    ('Dover', 'isA', 'City'),
+    ('Tallahassee', 'isA', 'City'),
+    ('Honolulu', 'isA', 'City'),
+    ('Boise', 'isA', 'City'),
+    ('Springfield', 'isA', 'City'),
+    ('Des Moines', 'isA', 'City'),
+    ('Topeka', 'isA', 'City'),
+    ('Frankfort', 'isA', 'City'),
+    ('Baton Rouge', 'isA', 'City'),
+    ('Augusta', 'isA', 'City'),
+    ('Annapolis', 'isA', 'City'),
+    ('Lansing', 'isA', 'City'),
+    ('Saint Paul', 'isA', 'City'),
+    ('Jackson', 'isA', 'City'),
+    ('Jefferson City', 'isA', 'City'),
+    ('Helena', 'isA', 'City'),
+    ('Lincoln', 'isA', 'City'),
+    ('Carson City', 'isA', 'City'),
+    ('Concord', 'isA', 'City'),
+    ('Trenton', 'isA', 'City'),
+    ('Santa Fe', 'isA', 'City'),
+    ('Albany', 'isA', 'City'),
+    ('Raleigh', 'isA', 'City'),
+    ('Bismarck', 'isA', 'City'),
+    ('Columbus', 'isA', 'City'),
+    ('Salem', 'isA', 'City'),
+    ('Harrisburg', 'isA', 'City'),
+    ('Providence', 'isA', 'City'),
+    ('Columbia', 'isA', 'City'),
+    ('Pierre', 'isA', 'City'),
+    ('Nashville', 'isA', 'City'),
+    ('Austin', 'isA', 'City'),
+    ('Montpelier', 'isA', 'City'),
+    ('Richmond', 'isA', 'City'),
+    ('Charleston', 'isA', 'City'),
+    ('Madison', 'isA', 'City'),
+    ('Cheyenne', 'isA', 'City'),
+    ('Charlotte Amalie', 'isA', 'City'),
+}
+
+is_country = {
+    ('Alabama', 'isA', 'state'),
+    ('Alaska', 'isA', 'state'),
+    ('Arizona', 'isA', 'state'),
+    ('Arkansas', 'isA', 'state'),
+    ('California', 'isA', 'state'),
+    ('Colorado', 'isA', 'state'),
+    ('Connecticut', 'isA', 'state'),
+    ('Delaware', 'isA', 'state'),
+    ('Florida', 'isA', 'state'),
+    ('Hawaii', 'isA', 'state'),
+    ('Idaho', 'isA', 'state'),
+    ('Illinois', 'isA', 'state'),
+    ('Iowa', 'isA', 'state'),
+    ('Kansas', 'isA', 'state'),
+    ('Kentucky', 'isA', 'state'),
+    ('Louisiana', 'isA', 'state'),
+    ('Maine', 'isA', 'state'),
+    ('Maryland', 'isA', 'state'),
+    ('Michigan', 'isA', 'state'),
+    ('Minnesota', 'isA', 'state'),
+    ('Mississippi', 'isA', 'state'),
+    ('Missouri', 'isA', 'state'),
+    ('Montana', 'isA', 'state'),
+    ('Nebraska', 'isA', 'state'),
+    ('Nevada', 'isA', 'state'),
+    ('New Hampshire', 'isA', 'state'),
+    ('New Jersey', 'isA', 'state'),
+    ('New Mexico', 'isA', 'state'),
+    ('New York', 'isA', 'state'),
+    ('North Carolina', 'isA', 'state'),
+    ('North Dakota', 'isA', 'state'),
+    ('Ohio', 'isA', 'state'),
+    ('Oregon', 'isA', 'state'),
+    ('Pennsylvania', 'isA', 'state'),
+    ('Rhode Island', 'isA', 'state'),
+    ('South Carolina', 'isA', 'state'),
+    ('South Dakota', 'isA', 'state'),
+    ('Tennessee', 'isA', 'state'),
+    ('Texas', 'isA', 'state'),
+    ('Vermont', 'isA', 'state'),
+    ('Virginia', 'isA', 'state'),
+    ('West Virginia', 'isA', 'state'),
+    ('Wisconsin', 'isA', 'state'),
+    ('Wyoming', 'isA', 'state'),
+    ('United States Virgin Islands', 'isA', 'state'),
+}
+
+# hasOptionally, hasCompulsory cannot be forced without controlling mechanism
+# which we do not want to implement. Some of the relations in the ontology
+# have only autodocumentation functionality
+#
+# Ontology variable should only to remind us that usually there are some handcrafted ontologies in the knowledge bases
+ontology = {
+    ('thing', 'hasProperty', 'hasProperty'),
+    ('state', 'hasProperty', 'isCapitalOf'),
+}
+
+plays_rock = {
+    ('ABBA', 'plays', 'rock'),
+    ('AC/DC', 'plays', 'rock'),
+    ('Paul Ackerman', 'plays', 'rock'),
+    ('Ad-Rock', 'plays', 'rock'),
+    ('Lou Adler', 'plays', 'rock'),
+    ('Steven Adler', 'plays', 'rock'),
+    ('Aerosmith', 'plays', 'rock'),
+    ('The Allman Brothers Band', 'plays', 'rock'),
+    ('Herb Alpert', 'plays', 'rock'),
+    ('The Animals', 'plays', 'rock'),
+    ('Louis Armstrong', 'plays', 'rock'),
+    ('Chet Atkins', 'plays', 'rock'),
+    ('LaVern Baker', 'plays', 'rock'),
+    ('Hank Ballard', 'plays', 'rock'),
+    ('The Band', 'plays', 'rock'),
+    ('Tony Banks (musician)', 'plays', 'rock'),
+    ('Syd Barrett', 'plays', 'rock'),
+    ('Frank Barsalona', 'plays', 'rock'),
+    ('Dave Bartholomew', 'plays', 'rock'),
+    ('Ralph Bass', 'plays', 'rock'),
+    ('The Beach Boys', 'plays', 'rock'),
+    ('Beastie Boys', 'plays', 'rock'),
+    ('The Beatles', 'plays', 'rock'),
+    ('Jeff Beck', 'plays', 'rock'),
+    ('Bee Gees', 'plays', 'rock'),
+    ('Benny Benjamin', 'plays', 'rock'),
+    ('Bobby Bennett (The Famous Flames)', 'plays', 'rock'),
+    ('Chuck Berry', 'plays', 'rock'),
+    ('Bill Haley & His Comets', 'plays', 'rock'),
+    ('Black Sabbath', 'plays', 'rock'),
+    ('Chris Blackwell', 'plays', 'rock'),
+    ('Otis Blackwell', 'plays', 'rock'),
+    ('Hal Blaine', 'plays', 'rock'),
+    ('Bobby Bland', 'plays', 'rock'),
+    ('Blondie', 'plays', 'rock'),
+    ('David Bowie', 'plays', 'rock'),
+    ('Charles Brown', 'plays', 'rock'),
+    ('James Brown', 'plays', 'rock'),
+    ('Ruth Brown', 'plays', 'rock'),
+    ('Jackson Browne', 'plays', 'rock'),
+    ('Michael Owen Bruce', 'plays', 'rock'),
+    ('Buffalo Springfield', 'plays', 'rock'),
+    ('Eric Burdon', 'plays', 'rock'),
+    ('Solomon Burke', 'plays', 'rock'),
+    ('Cliff Burton', 'plays', 'rock'),
+    ('James Burton', 'plays', 'rock'),
+    ('Geezer Butler', 'plays', 'rock'),
+    ('Jerry Butler', 'plays', 'rock'),
+    ('Paul Butterfield', 'plays', 'rock'),
+    ('Glen Buxton', 'plays', 'rock'),
+    ('Bobby Byrd', 'plays', 'rock'),
+    ('The Byrds', 'plays', 'rock'),
+    ('Jim Capaldi', 'plays', 'rock'),
+    ('Johnny Cash', 'plays', 'rock'),
+    ('Danny Cedrone', 'plays', 'rock'),
+    ('Ray Charles', 'plays', 'rock'),
+    ('Charlie Christian', 'plays', 'rock'),
+    ('Chuck D', 'plays', 'rock'),
+    ('Eric Clapton', 'plays', 'rock'),
+    ('Dick Clark', 'plays', 'rock'),
+    ('The Clash', 'plays', 'rock'),
+    ('Jimmy Cliff', 'plays', 'rock'),
+    ('The Coasters', 'plays', 'rock'),
+    ('Kurt Cobain', 'plays', 'rock'),
+    ('Eddie Cochran', 'plays', 'rock'),
+    ('Leonard Cohen', 'plays', 'rock'),
+    ('Nat King Cole', 'plays', 'rock'),
+    ('Phil Collins', 'plays', 'rock'),
+    ('Sam Cooke', 'plays', 'rock'),
+    ('Alice Cooper', 'plays', 'rock'),
+    ('Stewart Copeland', 'plays', 'rock'),
+    ('Elvis Costello', 'plays', 'rock'),
+    ('Floyd Cramer', 'plays', 'rock'),
+    ('Cream (band)', 'plays', 'rock'),
+    ('Creedence Clearwater Revival', 'plays', 'rock'),
+    ('Peter Criss', 'plays', 'rock'),
+    ('Crosby, Stills, Nash & Young', 'plays', 'rock'),
+    ('King Curtis', 'plays', 'rock'),
+    ('Roger Daltrey', 'plays', 'rock'),
+    ('Bobby Darin', 'plays', 'rock'),
+    ('The Dave Clark Five', 'plays', 'rock'),
+    ('Clive Davis', 'plays', 'rock'),
+    ('Miles Davis', 'plays', 'rock'),
+    ('The Dells', 'plays', 'rock'),
+    ('Mike D', 'plays', 'rock'),
+    ('Neil Diamond', 'plays', 'rock'),
+    ('Bo Diddley', 'plays', 'rock'),
+    ('Dion DiMucci', 'plays', 'rock'),
+    ('Willie Dixon', 'plays', 'rock'),
+    ('Fats Domino', 'plays', 'rock'),
+    ('Donovan', 'plays', 'rock'),
+    ('The Doors', 'plays', 'rock'),
+    ('Steve Douglas (saxophonist)', 'plays', 'rock'),
+    ('The Drifters', 'plays', 'rock'),
+    ('Dennis Dunaway', 'plays', 'rock'),
+    ('Bob Dylan', 'plays', 'rock'),
+    ('E Street Band', 'plays', 'rock'),
+    ('Eagles (band)', 'plays', 'rock'),
+    ('Earth, Wind & Fire', 'plays', 'rock'),
+    ('Duane Eddy', 'plays', 'rock'),
+    ('Ahmet Ertegün', 'plays', 'rock'),
+    ('Nesuhi Ertegun', 'plays', 'rock'),
+    ('The Everly Brothers', 'plays', 'rock'),
+    ('Faces (band)', 'plays', 'rock'),
+    ('The Famous Flames', 'plays', 'rock'),
+    ('Leo Fender', 'plays', 'rock'),
+    ('The Flamingos', 'plays', 'rock'),
+    ('Fleetwood Mac', 'plays', 'rock'),
+    ('The Four Seasons (band)', 'plays', 'rock'),
+    ('The Four Seasons Story', 'plays', 'rock'),
+    ('Four Tops', 'plays', 'rock'),
+    ('Aretha Franklin', 'plays', 'rock'),
+    ('Alan Freed', 'plays', 'rock'),
+    ('Ace Frehley', 'plays', 'rock'),
+    ('Milt Gabler', 'plays', 'rock'),
+    ('Peter Gabriel', 'plays', 'rock'),
+    ('Gamble and Huff', 'plays', 'rock'),
+    ('Carl Gardner', 'plays', 'rock'),
+    ('Marvin Gaye', 'plays', 'rock'),
+    ('Genesis (band)', 'plays', 'rock'),
+    ('Billy Gibbons', 'plays', 'rock'),
+    ('David Gilmour', 'plays', 'rock'),
+    ('Gerry Goffin', 'plays', 'rock'),
+    ('Berry Gordy', 'plays', 'rock'),
+    ('Bill Graham (promoter)', 'plays', 'rock'),
+    ('Grandmaster Flash and the Furious Five', 'plays', 'rock'),
+    ('Grateful Dead', 'plays', 'rock'),
+    ('Green Day', 'plays', 'rock'),
+    ('Al Green', 'plays', 'rock'),
+    ('Dave Grohl', 'plays', 'rock'),
+    ('Guns N Roses', 'plays', 'rock'),
+    ('G cont.', 'plays', 'rock'),
+    ('Woody Guthrie', 'plays', 'rock'),
+    ('Buddy Guy', 'plays', 'rock'),
+    ('Steve Hackett', 'plays', 'rock'),
+    ('Bill Haley', 'plays', 'rock'),
+    ('Hall & Oates', 'plays', 'rock'),
+    ('Kirk Hammett', 'plays', 'rock'),
+    ('George Harrison', 'plays', 'rock'),
+    ('Fuzzy Haskins', 'plays', 'rock'),
+    ('Isaac Hayes', 'plays', 'rock'),
+    ('Heart (band)', 'plays', 'rock'),
+    ('The Jimi Hendrix Experience', 'plays', 'rock'),
+    ('Jimi Hendrix', 'plays', 'rock'),
+    ('James Hetfield', 'plays', 'rock'),
+    ('Dusty Hill', 'plays', 'rock'),
+    ('Billie Holiday', 'plays', 'rock'),
+    ('Holland–Dozier–Holland', 'plays', 'rock'),
+    ('The Hollies', 'plays', 'rock'),
+    ('Buddy Holly', 'plays', 'rock'),
+    ('Jac Holzman', 'plays', 'rock'),
+    ('John Lee Hooker', 'plays', 'rock'),
+    ('Rick Huxley', 'plays', 'rock'),
+    ('The Impressions', 'plays', 'rock'),
+    ('The Ink Spots', 'plays', 'rock'),
+    ('Tony Iommi', 'plays', 'rock'),
+    ('The Isley Brothers', 'plays', 'rock'),
+    ('The Jackson 5', 'plays', 'rock'),
+    ('Mahalia Jackson', 'plays', 'rock'),
+    ('Michael Jackson', 'plays', 'rock'),
+    ('Wanda Jackson', 'plays', 'rock'),
+    ('James Jamerson', 'plays', 'rock'),
+    ('Etta James', 'plays', 'rock'),
+    ('Jefferson Airplane', 'plays', 'rock'),
+    ('Billy Joel', 'plays', 'rock'),
+    ('Dr. John', 'plays', 'rock'),
+    ('Elton John', 'plays', 'rock'),
+    ('Little Willie John', 'plays', 'rock'),
+    ('Brian Johnson', 'plays', 'rock'),
+    ('Johnnie Johnson (musician)', 'plays', 'rock'),
+    ('Robert Johnson', 'plays', 'rock'),
+    ('Quincy Jones', 'plays', 'rock'),
+    ('Janis Joplin', 'plays', 'rock'),
+    ('Louis Jordan', 'plays', 'rock'),
+    ('Albert King', 'plays', 'rock'),
+    ('B.B. King', 'plays', 'rock'),
+    ('Carole King', 'plays', 'rock'),
+    ('Freddie King', 'plays', 'rock'),
+    ('The Kinks', 'plays', 'rock'),
+    ('Kiss (band)', 'plays', 'rock'),
+    ('Gladys Knight & the Pips', 'plays', 'rock'),
+    ('Ronnie Lane', 'plays', 'rock'),
+    ('Lead Belly', 'plays', 'rock'),
+    ('Led Zeppelin', 'plays', 'rock'),
+    ('Brenda Lee', 'plays', 'rock'),
+    ('Geddy Lee', 'plays', 'rock'),
+    ('Jerry Leiber and Mike Stoller', 'plays', 'rock'),
+    ('John Lennon', 'plays', 'rock'),
+    ('Jerry Lee Lewis', 'plays', 'rock'),
+    ('Alex Lifeson', 'plays', 'rock'),
+    ('List of awards and nominations received by The Who', 'plays', 'rock'),
+    ('Little Anthony and the Imperials', 'plays', 'rock'),
+    ('Little Richard', 'plays', 'rock'),
+    ('Little Walter', 'plays', 'rock'),
+    ('Vini Lopez', 'plays', 'rock'),
+    ('Darlene Love', 'plays', 'rock'),
+}
+
+data = is_capital | is_city | is_country | ontology | plays_rock
